+++ src/components/jsNMR/src/nmr.js
@@ -82,7 +82,8 @@
 
 			for( var i = 0, l = nmr.integrals[ mode ].length; i < l ; i ++ ) {
 
-				nmr.integrals[ mode ][ i ].ratio = sumMax == 0 ? 1 : nmr.integrals[ mode ][ i ].lastSum / sumMax;
+				nmr.integrals[ mode ][ i ].ratio = Math.abs( sumMax == 0 ? 1 : nmr.integrals[ mode ][ i ].lastSum / sumMax );
+
 				nmr.integrals[ mode ][ i ].setPosition();
 
 				if( nmr.integralBasis ) {
@@ -498,6 +499,17 @@
 			return this.options.dom;
 		}
 
+		NMR.prototype.getGraph2D = function() {
+			return this.graphs['_2d'];
+		}
+
+		NMR.prototype.getGraphX = function() {
+			return this.graphs['x'];
+		}
+
+		NMR.prototype.getGraphY = function() {
+			return this.graphs['y'];
+		}
 
 		NMR.prototype.resize1DTo = function( w, h ) {
 			this.graphs[ 'x' ].resize( w, h );
@@ -567,7 +579,7 @@
 			}
 
 			var serie_y = this.graphs['y']
-				.newSerie(name, $.extend( { useSlots: true }, options ) )
+				.newSerie(name, $.extend( { useSlots: true, flip: true }, options ) )
 				.setLabel( "My serie" )
 				.setXAxis( this.graphs['y'].getBottomAxis( ) )
 				.setYAxis( this.graphs['y'].getRightAxis( ) )
@@ -681,6 +693,11 @@
 			this.graphs[ 'y' ].updateAxes();
 			this.graphs[ 'x' ].updateAxes();
 
+			var yaxis = this.graphs[ 'y' ].getXAxis();
+			var xaxis = this.graphs[ 'x' ].getYAxis();
+
+			this.graphs[ '_2d' ].getXAxis().force( xaxis );
+			this.graphs[ '_2d' ].getYAxis().force( yaxis );
 
 			this.graphs['y'].redraw( );
 			this.graphs['y'].drawSeries();	
@@ -854,8 +871,6 @@
 
 						} );
 
-						// ANDRES
-						// You can do here your processing and create new shapes
 						shape.kill();
 					}
 				},
@@ -877,8 +892,10 @@
 						},
 
 						onZoomEnd: function( graph, x, y, e, target, x1, y1 ) {
-							self.graphs['x']._pluginExecute( 'graph.plugin.zoom', 'onMouseUp', [ self.graphs['x'], x, y, e, true ] );
-							self.graphs['y']._pluginExecute( 'graph.plugin.zoom', 'onMouseUp', [ self.graphs['y'], x, y, e, true ] );
+
+ 							// Remove shapes
+							self.graphs[ 'x' ]._pluginExecute( 'graph.plugin.zoom', 'removeZone', [ ] );
+							self.graphs[ 'y' ]._pluginExecute( 'graph.plugin.zoom', 'removeZone', [ ] );
 
 							var x = graph.getBottomAxis().getVal( x1 );
 							var x2 = graph.getBottomAxis().getVal( x );
@@ -886,10 +903,21 @@
 							var y = graph.getLeftAxis().getVal( y1 );
 							var y2 = graph.getLeftAxis().getVal( y );
 
+							// Resize x
+ 							self.graphs['x']._applyToAxes( '_doZoomVal', [ x, x2 ], false, true );
+ 							self.graphs['x'].redraw();
+ 							self.graphs['x'].drawSeries();
+
+
+							// Resize y
+ 							self.graphs['x']._applyToAxes( '_doZoomVal', [ y, y2 ], false, true );
+ 							self.graphs['x'].redraw();
+ 							self.graphs['x'].drawSeries();
+
 							if( self.options.minimap ) {
+
 								self.minimapClip.data.pos = { x: x, y: y };
 								self.minimapClip.data.pos2 = { x: x2, y: y2 };
-
 								self.minimapClip.redraw();
 							}
 						},
@@ -1053,6 +1081,7 @@
 			self.graphs[ '_2d' ].getXAxis().setAxisDataSpacing( 0 );
 			self.graphs[ '_2d' ].getYAxis().setAxisDataSpacing( 0 );
 
+
 	
 			var legend = this.graphs[ '_2d' ].makeLegend( { frame: true, frameColor: 'grey', frameWidth: 1, movable: true } );
 			legend.setPosition( { x: '20px', y: '20px' } );
@@ -1129,7 +1158,15 @@
 
 						onZoomEnd: function( graph, x, y, e, target ) {
 
-							self.graphs[ '_2d' ]._pluginExecute( 'graph.plugin.zoom', 'onMouseUp', [ self.graphs[ '_2d' ], x, undefined, e, true ] );
+
+							var xaxis = self.graphs['x'].getXAxis();
+							var from = xaxis.getActualMin();
+							var to = xaxis.getActualMax();
+
+ 							self.graphs['_2d']._applyToAxes( '_doZoomVal', [ from, to ], true, false );
+ 							self.graphs['_2d'].redraw();
+ 							self.graphs['_2d'].drawSeries();
+							self.graphs[ '_2d' ]._pluginExecute( 'graph.plugin.zoom', 'removeZone', [ ] );
 
 						},
 
@@ -1163,7 +1200,8 @@
 					type: 'plugin',
 					plugin: 'graph.plugin.zoom',
 					options: {
-						direction: 'x'
+						direction: 'y',
+						baseline: 0
 					}
 				},
 
@@ -1187,6 +1225,8 @@
 			  	}
 
 			} );
+
+		//	self.graphs[ 'x' ].getXAxis().options.wheelBaseline = 0;
 
 
 
@@ -1213,7 +1253,26 @@
 
 						onZoomEnd: function( graph, x, y, e, target ) {
 
-							self.graphs[ '_2d' ]._pluginExecute( 'graph.plugin.zoom', 'onMouseUp', [ self.graphs[ '_2d' ], undefined, y, e, true ] );
+							var yaxis = self.graphs['y'].getYAxis();
+							var from = yaxis.getActualMin();
+							var to = yaxis.getActualMax();
+
+ 							self.graphs['_2d']._applyToAxes( '_doZoomVal', [ from, to ], false, true );
+ 							self.graphs['_2d'].redraw();
+ 							self.graphs['_2d'].drawSeries();
+							self.graphs[ '_2d' ]._pluginExecute( 'graph.plugin.zoom', 'removeZone', [ ] );
+
+/*
+
+							var xaxis = self.graphs['y'].getYAxis();
+							var from = xaxis.getCurrentMin();
+							var to = xaxis.getCurrentMax();
+console.log( from, to );
+ 							self.graphs['_2d']._applyToAxes( '_doZoomVal', [ from, to ], false, true );
+
+ 							*/
+
+							//self.graphs[ '_2d' ]._pluginExecute( 'graph.plugin.zoom', 'onMouseUp', [ self.graphs[ '_2d' ], undefined, y, e, true ] );
 
 						},
 
@@ -1245,7 +1304,8 @@
 					type: 'plugin',
 					plugin: 'graph.plugin.zoom',
 					options: {
-						direction: 'y'
+						direction: 'y',
+						baseline: 0
 					}
 				},
 
@@ -1330,7 +1390,8 @@
 					type: 'plugin',
 					plugin: 'graph.plugin.zoom',
 					options: {
-						direction: 'x'
+						direction: 'y',
+						baseline: 0
 					}
 				},
 
@@ -1351,7 +1412,8 @@
 
 
 			this.graphs[ 'x' ].setHeight(300);
-
+	//		this.graphs[ 'x' ].getXAxis().options.wheelBaseline = 0;
+			
 			this.graphs[ 'x' ].shapeHandlers.onRemoved.push( function( shape ) {