+++ src/init.js
@@ -17,7 +17,7 @@ requirejs.config({
 		ChemDoodle:         'lib/chemdoodle/ChemDoodleWeb-unpacked',
 		pouchdb:            'components/pouchdb/dist/pouchdb.min',
 		uri:                'components/uri.js/src',
-		underscore:         'components/underscore/underscore',
+		underscore:         'components/underscore/underscore-min',
         lodash:             'components/lodash/dist/lodash.min',
         modernizr:          'components/modernizr/modernizr',
         bowser:             'components/bowser/bowser.min',
@@ -31,7 +31,11 @@ requirejs.config({
 			exports: 'd3'
 		},
 		threejs: {
-			exports: 'THREE'
+			exports: 'THREE',
+			init: function() {
+				// Allow cross-origin images
+				this.THREE.ImageUtils.crossOrigin = '';
+			}
 		},
 		'components/x2js/xml2json.min': {
 			exports: 'X2JS'