+++ src/modules/types/client_interaction/code_editor/view.js
@@ -37,6 +37,7 @@ define(['modules/default/defaultview', 'src/util/util', 'ace/ace', 'src/util/con
                 this.editor.getSession().on('change', function () {
                     self.editorChanged();
                 });
+                this.editorChanged();
             }
 
             if (this.module.getConfigurationCheckbox('hasButton', 'button')) {
@@ -67,7 +68,9 @@ define(['modules/default/defaultview', 'src/util/util', 'ace/ace', 'src/util/con
         editorChanged: function () {
             var val = this.editor.getValue();
             this._code = val;
-            this.module.definition.configuration.groups.group[0].script[0] = val;
+            if(this.module.definition.configuration.groups) {
+                this.module.definition.configuration.groups.group[0].script[0] = val;
+            }
             this.module.controller.onEditorChanged(this._code);
         }
     });